requirements.txt
Flask>=2.0
groq>=0.4.0
python-dotenv>=1.0.0
gunicorn
flake8
pytest
black
bandit
pyright
semgrep
pip-audit
pytest-cov
openai
google-generativeai